Get a glimpse of your Spat* objects

glimpse() is like a transposed version of print(): layers/columns run down the page, and data runs across. This makes it possible to see every layer/column in a Spat* object.

Usage

# S3 method for SpatRaster
glimpse(x, width = NULL, ...)

# S3 method for SpatVector
glimpse(x, width = NULL, ...)

Arguments

x

A SpatRaster created with terra::rast() or a SpatVector created with terra::vect().

width

Width of output: defaults to the setting of the width if finite (see dplyr::glimpse()) or the width of the console.

...

Arguments passed on to as_tibble() Spat methods.

Value

original x is (invisibly) returned, allowing glimpse() to be used within a data pipeline.

terra equivalent

print()

Methods

Implementation of the generic dplyr::glimpse() function for Spat*. objects.

See also

Examples

library(terra)

# SpatVector
v <- vect(system.file("extdata/cyl.gpkg", package = "tidyterra"))

v %>% glimpse(width = 200)
#> Rows: 9
#> Columns: 3
#> $ iso2 <chr> "ES-AV", "ES-BU", "ES-LE", "ES-P", "ES-SA", "ES-SG", "ES-SO", "ES-VA", "ES-ZA"
#> $ cpro <chr> "05", "09", "24", "34", "37", "40", "42", "47", "49"
#> $ name <chr> "Avila", "Burgos", "Leon", "Palencia", "Salamanca", "Segovia", "Soria", "Valladolid", "Zamora"

# Use on a pipeline
v %>%
  glimpse() %>%
  mutate(a = 30) %>%
  # with options
  glimpse(geom = "WKT")
#> Rows: 9
#> Columns: 3
#> $ iso2 <chr> "ES-AV", "ES-BU", "ES-LE", "ES-P", "ES-SA", "ES-SG", "ES-SO", "ES…
#> $ cpro <chr> "05", "09", "24", "34", "37", "40", "42", "47", "49"
#> $ name <chr> "Avila", "Burgos", "Leon", "Palencia", "Salamanca", "Segovia", "S…
#> Rows: 9
#> Columns: 5
#> $ iso2     <chr> "ES-AV", "ES-BU", "ES-LE", "ES-P", "ES-SA", "ES-SG", "ES-SO",…
#> $ cpro     <chr> "05", "09", "24", "34", "37", "40", "42", "47", "49"
#> $ name     <chr> "Avila", "Burgos", "Leon", "Palencia", "Salamanca", "Segovia"…
#> $ a        <dbl> 30, 30, 30, 30, 30, 30, 30, 30, 30
#> $ geometry <chr> "POLYGON ((3126360.2417 2066777.7545, 3125073.9752 2065007.29…

# SpatVector
r <- rast(system.file("extdata/cyl_elev.tif", package = "tidyterra"))

r %>% glimpse()
#> Rows: 26,712
#> Columns: 1
#> $ elevation_m <dbl> 700.2969, 780.3889, 706.1250, 568.9722, 584.9028, 506.7361…

# Use on a pipeline
r %>%
  glimpse() %>%
  mutate(b = elevation_m / 100) %>%
  # With options
  glimpse(xy = TRUE)
#> Rows: 26,712
#> Columns: 1
#> $ elevation_m <dbl> 700.2969, 780.3889, 706.1250, 568.9722, 584.9028, 506.7361…
#> Rows: 26,712
#> Columns: 4
#> $ x           <dbl> -7.0625, -7.0375, -7.0125, -6.9875, -6.9625, -6.9375, -6.9…
#> $ y           <dbl> 43.22083, 43.22083, 43.22083, 43.22083, 43.22083, 43.22083…
#> $ elevation_m <dbl> 700.2969, 780.3889, 706.1250, 568.9722, 584.9028, 506.7361…
#> $ b           <dbl> 7.002969, 7.803889, 7.061250, 5.689722, 5.849028, 5.067361…